use `qemu_test_runner_toolchain` to run `cc_test` targets in QEMU
Add a CcTestRunnerToolchain that is used to run `cc_test` targets in
QEMU. When using this CcTestRunnerToolchain, semihosting must be enabled
-- it is not automatically enabled for building the `cc_test` target.

This is because the configuration must by applied to the underlying
`lm3s6965evb` toolchain (to select startup code that enables
semihosting) which is selected by toolchain registration when using
`--platforms=//platform:lm3s6965evb`.

The underlying QEMU runner of the test runner toolchain
(`//:qemu_runner`) always uses semihosting, as it would be meaning less
otherwise -- the binaries must complete and test success is determined
by the return code.

```sh
bazel test \
  --platforms=//platform:lm3s6965evb \
  --extra_toolchains=//:qemu_test_runner_toolchain \
  --//config:semihosting=true \
  //:dummy_test
```

Change-Id: I4b78fe92fd5cef08271223ede19c44d0f94b18d4