define `transition_config_{binary,test}` rules (#18)
`transition_config_binary` and `transition_config_test` are used to
transition a `cc_binary` or `cc_test` target to a specified platform,
removing the need to specify the platform at the command line. With
`cc_binary`, the corresponding binary still need to be run with
`--run_under=//:qemu_runner` if the target platform differs from the
host platform -- this is because `--run_under` depends on the exec
configuration and is not a valid build setting for transitions of a
target.

`transition_config_{binary,test}` have the following attributes:
`src`: `label`
The binary target to transition.
  
`inherit_env`: `string_list`
Environment variables to inherit when running the transitioned binary.

`semihosting`: `string | None`
 Enable or disable semihosting, or leave unchanged if `None`.
  
`platform`: `label`
 Target platform for the binary.

`extra_toolchains`: `label_list`
 Extra toolchains to use when building the binary. May also include
 extra toolchains to use when running the binary if it is a test.

`semihosting` defaults to `None` for `transition_config_binary`
and `"enabled"` for `transition_config_test`.

Note that the `//:qemu_test_runner_toolchain` can be set in
`extra_toolchains` for test targets as an alternative, or to override,
registered toolchains.

The QEMU machine is encoded in the transitioned target via
RunEnvironmentInfo, which is then read by the QemuRunner.

These rules allow flagless builds for building and testing (but not
running):

```starlark
transition_config_binary(
    name = "my_binary.lm3s6965evb",
    src = ":my_binary",
    platform = "//platform:lm3s6965evb",
    semihosting = "enabled",
)

transition_config_test(
    name = "my_test.lm3s6965evb",
    src = ":my_test",
    platform = "//platform:lm3s6965evb",
)
```

Change-Id: Id53512383255bc23e5f3216b9bd39708687f3d3e