define a `gdb_qemu_runner`
Provide a new runner that runs a binary with QEMU as a background
process and then connects to it with GDB.

This simplifies the debug process as the user can use a single `bazel
run` command instead of two.

From the example directory:

  bazel run \
    --run_under=@cortex_m//:gdb_qemu_runner \
    -c dbg \
    //semihosting:semihosting.lm3s6965evb

or from the root directory:

  bazel run \
    --run_under=//:gdb_qemu_runner \
    -c dbg \
    //test:hello_semihosting_test.transition

By default the `//:gdb_qemu_runner` will launch QEMU with semihosting
enabled, start a GDB server on port 1234, and freeze the CPU at startup.
GDB is then started and will connect to QEMU.

Change-Id: I05fa57a243ca6f46b5186cae58c3a80031029536