run bazel within a nix shell
Add a `tools/bazel` wrapper to launch the Bazel process within a Nix
shell. This shell establishes a PATH only using the Nix store to ensure
that all inputs of the build are specified.

Note that this shell is only used when building the `cortex_m` repo --
not when using `cortex_m` as an external.

In order to start the Nix shell, which then calls `bazelisk` from Nix,
`bash` and `nix` must be in the PATH.

https://github.com/bazelbuild/bazelisk?tab=readme-ov-file#ensuring-that-your-developers-use-bazelisk-rather-than-bazel

Change-Id: Iefaed1786aed904215cb5fe3e1135b4a56f6d14f