use a hermetic shell with `tools/bazel` (#51)
Add a `tools/bazel` [wrapper] to launch the Bazel process within a Nix
shell application . This shell establishes a PATH only using the Nix
store to ensure that all inputs of the build are specified.

Note that this shell is only used when developing the cortex_m repo --
not when using cortex_m as an external.

In order to bootstrap the Nix shell application, which then calls
`bazelisk` from Nix, `bash`, `nix`, and coreutils must be in the PATH.
Bootstrapping doesn't use `nix shell` due to the large overhead (~1
second on my machine). Instead the shell application derivation is
"compiled" to avoid re-derivation overhead. This approach is similar to
[nix-script].

[wrapper]:
https://github.com/bazelbuild/bazelisk?tab=readme-ov-file#toolsbazel
[nix-script]: https://bytes.zone/posts/nix-script/

Change-Id: Iefaed1786aed904215cb5fe3e1135b4a56f6d14f