add tooling for Nix files
Define a `flake_package_dev_deps` extension that obtains development
dependencies from nixpkgs. This uses the existing `flake.nix` in
`//extensions`. The dev dependencies include:
* `nixd`
* `nixfmt`
* `nixfmt-tree`

`nixd` (and `nixfmt`) must be fetched explicitly with
```
bazel fetch @nixd
```
as these are only used as part of the Zed project settings.

This PR also adds `gnused` as a package, as it is used extensively by
shell runfiles:
https://github.com/bazelbuild/rules_shell/blob/main/shell/runfiles/runfiles.bash#L184

Change-Id: Ibb8c58fbf876a2eb3c612b3985f45840be7f7d79