use glibc sysroot from nixpkgs
Provide libxml2 and libz as shared library dependencies for clang
binaries (e.g. clang, lld).

Configure the LLVM toolchain to use glibc from nixpkgs as the sysroot.
nixpkgs provides libc.so and libm.so as linker scripts so we remove
these from the exported glibc package and link directly against
`libc.so.6` and `libm.so.6`.