use `glibc` sysroot for linux host machines
On Linux, provide a glibc sysroot from Nixpkgs instead of relying on the
host system to provide one. The LLVM toolchain is also configured to use
this sysroot.

When fetching `@toolchains_llvm`, the shebang is patched for the
`cc_wrapper` script used for Linux, allowing use with NixOS.

The `bazel-wrapper` script now provides `libxml2` and `libz` so that
they can be used by the LLVM toolchain binaries (e.g clang, lld), as
these have shared library dependencies.

Change-Id: Idd176574d437acc34c574ea5d3f5fc820316963a